src/main/java/hello2021/hellospring2021/repository/MemoryMemberRepository.java
package hello2021.hellospring2021.repository;

import hello2021.hellospring2021.domain.Member;
import org.springframework.stereotype.Repository;

import java.util.*;


public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long,Member> store = new HashMap<>();
    private static long sequence=0L;

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(),member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id)); //null반환시 사용
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream() //람다식 store...멤버에 있는 값과 파라메터 name을 맞춰 찾으면 반환.
                .filter(member -> member.getName().equals(name))
                .findAny();

    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore(){
        store.clear();
    }
}
